# target/classes/api/api.yml
openapi: 3.0.1
info:
  title: Product API
  description: API pour gérer les produits
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:

  /internal/api/v1/products:
    get:
      tags:
        - "Products"
      operationId: "getAllProducts"
      summary: Récupérer tous les produits
      responses:
        '200':
          description: Succès. Renvoie la liste de tous les produits.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Mauvaise requête
        '404':
          description: Produits non retrouvables
        '500':
          description: Erreur interne du serveur

    post:
      tags:
        - "Products"
      operationId: "createProduct"
      summary: Ajoute un nouveau produit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'

      responses:
        '201':
          description: Produit ajouté avec succès
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Mauvaise requête.
        '500':
          description: Erreur interne du serveur

  /internal/api/v1/products/{id}:
    get:
      tags:
        - "Products"
      operationId: "getProductById"
      parameters:
        - name: id
          in: path
          description: ID du produit à retrouver
          required: true
          schema:
            type: integer
            format: int64
      summary: Récupérer un produit à partir de son identifiant
      responses:
        '200':
          description: Succès. Renvoie le produit ayant l'identifiant concerné.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: ID invalide
        '404':
          description: Produit non retrouvable

    put:
      tags:
        - "Products"
      operationId: "updateProduct"
      summary: Modifie un produit
      parameters:
        - name: id
          in: path
          description: ID du produit à retrouver
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Produit modifié avec succès
        '400':
          description: Mauvaise requête.
        '500':
          description: Erreur interne du serveur

    delete:
      tags:
        - "Products"
      operationId: "deleteProductById"
      parameters:
        - name: id
          in: path
          description: ID du produit à supprimer
          required: true
          schema:
            type: integer
            format: int64
      summary: Supprimer un produit à partir de son identifiant
      responses:
        '400':
          description: ID invalide

components:
  schemas:
    Product:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        price:
          type: string
        description:
          type: string
        color:
          type: string
        stock:
          type: integer
          format: int64